JunctionTrees

Trees

StructuredDecompositions.JunctionTrees.Tree — Type

Tree <: AbstractUnitRange{Int}

A rooted forest. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.AbstractTree — Type

AbstractTree = Union{Tree, SupernodeTree, JunctionTree}

A rooted forest. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.eliminationtree — Function

eliminationtree(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

A non-mutating version of eliminationtree!.

StructuredDecompositions.JunctionTrees.eliminationtree! — Function

eliminationtree!(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a tree-depth decomposition of a simple graph. See junctiontree! for the meaning of alg.

julia> using StructuredDecompositions

julia> graph = [
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ];

julia> label, tree = eliminationtree(graph);

julia> tree
8-element Tree:
8
└─ 7
   ├─ 5
   │  ├─ 1
   │  └─ 4
   │     └─ 3
   │        └─ 2
   └─ 6

AbstractTrees.rootindex — Function

rootindex(tree::AbstractTree)

Get the least root of a rooted forest.

AbstractTrees.parentindex — Function

parentindex(tree::AbstractTree)

Get the parent index of node i. Returns nothing if i is a root.

StructuredDecompositions.JunctionTrees.firstchildindex — Function

firstchildindex(tree::AbstractTree, i::Integer)

Get the least child of node i. Returns nothing if i is a leaf.

AbstractTrees.nextsiblingindex — Function

nextsiblingindex(tree::AbstractTree, i::Integer)

Get the next sibling of node i. Returns nothing if i is the greatest sibling.

StructuredDecompositions.JunctionTrees.rootindices — Function

rootindices(tree::AbstractTree)

Construct an iterator over the roots of a rooted forest.

AbstractTrees.childindices — Function

childindices(tree::AbstractTree, i::Integer)

Construct an iterator over the children of node i.

Junction Trees

StructuredDecompositions.JunctionTrees.Bag — Type

Bag <: AbstractVector{Int}

A bag of a junction tree.

StructuredDecompositions.JunctionTrees.JunctionTree — Type

JunctionTree <: AbstractVector{Bag}

A junction tree. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.junctiontree! — Function

junctiontree!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

A mutating version of junctiontree.

StructuredDecompositions.JunctionTrees.junctiontree — Function

junctiontree(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a tree decomposition of a simple graph. The vertices of the graph are first ordered by a fill-reducing permutation computed by the algorithm alg. The size of the resulting decomposition is determined by the supernode partition snd.

julia> using StructuredDecompositions

julia> graph = [
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ];

julia> label, tree = junctiontree(graph);

julia> tree
6-element JunctionTree:
[6, 7, 8]
├─ [1, 6, 7]
├─ [4, 6, 8]
│  └─ [3, 4, 6]
│     └─ [2, 3, 6]
└─ [5, 7, 8]

StructuredDecompositions.JunctionTrees.treewidth! — Function

treewidth!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

A mutating version of treewidth.

StructuredDecompositions.JunctionTrees.treewidth — Function

treewidth(tree::JunctionTree)

Compute the width of a junction tree.

treewidth(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Compute an upper bound to the tree width of a simple graph. See junctiontree for the meaning of alg.

StructuredDecompositions.JunctionTrees.separator — Function

separator(bag::Bag)

Get the separator of a bag.

separator(tree::JunctionTree, i::Integer)

Get the separator at node i.

StructuredDecompositions.JunctionTrees.residual — Function

residual(bag::Bag)

Get the residual of a bag.

residual(tree::JunctionTree, i::Integer)

Get the residual at node i.

StructuredDecompositions.JunctionTrees.relative — Function

relative(tree::JunctionTree, i::Integer)

Get the indices in tree[parentindex(tree, i)] corresponding to the elements of separator(tree, i).

julia> using AbstractTrees

julia> using StructuredDecompositions

julia> graph = [
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ];

julia> label, tree = junctiontree(graph);

julia> bag = tree[parentindex(tree, 1)]
3-element Bag:
 6
 7
 8

julia> sep = separator(tree, 1)
2-element view(::Vector{Int64}, 1:2) with eltype Int64:
 6
 7

julia> rel = relative(tree, 1)
2-element view(::Vector{Int64}, 1:2) with eltype Int64:
 1
 2

julia> bag[rel] == sep
true

Chordal Graphs

StructuredDecompositions.JunctionTrees.eliminationgraph — Function

eliminationgraph([element=true,] tree::JunctionTree)

See [eliminationgraph!]. The function returns a sparse matrix whose structural nonzeros are filled with element.

eliminationgraph(T::Type, tree::JunctionTree)

See [eliminationgraph!]. The function returns a sparse matrix with elements of type T.

eliminationgraph(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct the elimination graph of a simple graph. ```julia julia> using StructuredDecompositions, SparseArrays

julia> graph = [ 0 1 1 0 0 0 0 0 1 0 1 0 0 1 0 0 1 1 0 1 1 0 0 0 0 0 1 0 1 0 0 0 0 0 1 1 0 0 1 1 0 1 0 0 0 0 1 0 0 0 0 0 1 1 0 1 0 0 0 0 1 0 1 0 ];

julia> label, filled = label, filled = eliminationgraph(graph);

julia> filled 8×8 SparseMatrixCSC{Int64, Int64} with 13 stored entries: ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ 1 ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ 1 ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ ⋅ 1 1 1 1 ⋅ ⋅ ⋅ ⋅ 1 ⋅ ⋅ ⋅ 1 1 ⋅ ⋅ ⋅ ⋅ ⋅ 1 1 1 1 ⋅

julia> isfilled(filled) true

julia> ischordal(filled + filled') true

StructuredDecompositions.JunctionTrees.eliminationgraph! — Function

eliminationgraph!(target::SparseMatrixCSC, tree::JunctionTree)

Construct the subtree graph of a junction tree. The result is stored in target.

eliminationgraph!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

A mutating version of eliminationgraph.

StructuredDecompositions.JunctionTrees.ischordal — Function

ischordal(matrix::AbstractMatrix)

Determine whether a simple graph is chordal.

StructuredDecompositions.JunctionTrees.isfilled — Function

isfilled(matrix::AbstractMatrix)

Determine whether a directed graph is filled.

StructuredDecompositions.JunctionTrees.isperfect — Function

isperfect(matrix::AbstractMatrix, order::AbstractVector[, index::AbstractVector])

Determine whether an fill-reducing permutation is perfect.

Elimination Algorithms

StructuredDecompositions.JunctionTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

type	name	complexity

MCS	maximum cardinality search	O(m + n)

RCM	reverse Cuthill-Mckee	O(mΔ)

AMD	approximate minimum degree	O(mn)

SymAMD	column approximate minimum degree	O(mn)

MMD	multiple minimum degree	O(mn²)

NodeND	nested dissection

Spectral	spectral ordering

BT	Bouchitte-Todinca	O(2.6183ⁿ)

for a graph with m edges, n vertices, and maximum degree Δ.

StructuredDecompositions.JunctionTrees.PermutationOrAlgorithm — Type

PermutationOrAlgorithm = Union{AbstractVector, EliminationAlgorithm}

Either a permutation or an algorithm.

StructuredDecompositions.JunctionTrees.DEFAULT_ELIMINATION_ALGORITHM — Constant

DEFAULT_ELIMINATION_ALGORITHM = AMD()

The default algorithm.

StructuredDecompositions.JunctionTrees.MCS — Type

MCS <: EliminationAlgorithm

MCS()

The maximum cardinality search algorithm.

StructuredDecompositions.JunctionTrees.RCM — Type

RCM <: EliminationAlgorithm

RCM(; sortbydeg=true)

The reverse Cuthill-McKee algorithm.

sortbydeg: whether to sort neighbor lists by degree

StructuredDecompositions.JunctionTrees.AMD — Type

AMD <: EliminationAlgorithm

AMD(; dense=10.0, aggressive=1.0)

The approximate minimum degree algorithm.

dense: dense row parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.SymAMD — Type

SymAMD <: EliminationAlgorithm

SymAMD(; dense_row=10.0, dense_col=10.0, aggressive=1.0)

The column approximate minimum degree algorithm.

dense_row: dense row parameter

dense_column: dense column parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.MMD — Type

MMD <: EliminationAlgorithm

MMD()

The multiple minimum degree algorithm.

StructuredDecompositions.JunctionTrees.NodeND — Type

NodeND <: EliminationAlgorithm

NodeND()

The nested dissection algorithm. In order to use this algorithm, import the package Metis.

StructuredDecompositions.JunctionTrees.Spectral — Type

Spectral <: EliminationAlgorithm

Spectral(; tol=0.0)

The spectral ordering algorithm. In order to use this algorithm, import the package Laplacians.

tol: tolerance for convergence

StructuredDecompositions.JunctionTrees.BT — Type

BT <: EliminationAlgorithm

BT()

The Bouchitte-Todinca algorithm. In order to use this algorithm, import the package TreeWidthSolver.

StructuredDecompositions.JunctionTrees.permutation — Function

permutation(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a fill-reducing permutation of the vertices of a simple graph.

julia> using StructuredDecompositions

julia> graph = [
    0 1 1 0 0 0 0 0
    1 0 1 0 0 1 0 0
    1 1 0 1 1 0 0 0
    0 0 1 0 1 0 0 0
    0 0 1 1 0 0 1 1
    0 1 0 0 0 0 1 0
    0 0 0 0 1 1 0 1
    0 0 0 0 1 0 1 0
];

julia> order, index = permutation(graph; alg=MCS());

julia> order
8-element Vector{Int64}:
 1
 6
 2
 3
 4
 5
 7
 8

julia> index == invperm(order)
true

Supernodes

StructuredDecompositions.JunctionTrees.SupernodeType — Type

SupernodeType

A type of supernode partition. The options are

type	name

Nodal	nodal supernode partition

Maximal	maximal supernode partition

Fundamental	fundamental supernode partition

StructuredDecompositions.JunctionTrees.DEFAULT_SUPERNODE_TYPE — Constant

DEFAULT_SUPERNODE_TYPE = Maximal()

The default supernode partition.

StructuredDecompositions.JunctionTrees.Nodal — Type

Nodal <: SupernodeType

A nodal supernode partition.

StructuredDecompositions.JunctionTrees.Maximal — Type

Maximal <: SupernodeType

A maximal supernode partition.

StructuredDecompositions.JunctionTrees.Fundamental — Type

Fundamental <: SupernodeType

A fundamental supernode partition.