Library Reference

StructuredDecompositions.Decompositions.StrDecomp — Type

StrDecomp(graph::AbstractSymmetricGraph[, ealg::Union{Order, EliminationAlgorithm}[, stype::SupernodeType]])

Construct a structured decomposition of a simple graph, optionally specifying an elimination algorithm and supernode type.

StructuredDecompositions.Decompositions.StrDecomp — Type

Structured decomposition struct – Consider these as graphs whose vertices are labeled by the objects of some category and whose edges are labeled by spans in this category

StructuredDecompositions.Decompositions.StrDecomp — Method

One can construct a structured decomposition by simply providing the graph representing the shape of the decompostion and the relevant diagram. This constructor will default to setting the Decompsotion Type to Decomposition (i.e. we default to viewing C-valued structured decompositions as diagrams into Span C)

StructuredDecompositions.Decompositions.StructuredDecomposition — Type

Structured decompositions are diagrams.

StructuredDecompositions.Decompositions.adhesionSpans — Method

Get a vector of indexed adhesion spans; i.e. a vector consisting of pairs (x₁ <- e -> x₂, dx₁ <- de -> dx₂) where x₁ <- e -> x₂ is span in ∫G and dx₁ <- de -> dx₂ is what its image under the decompositon d

StructuredDecompositions.Decompositions.adhesionSpans — Method

Get a vector of the adhesion spans of a decomposition

StructuredDecompositions.Decompositions.adhesions — Method

Get a vector of indexed adhesions; i.e. a vector consisting of pairs (e, de) where e is an edge in ∫G and de is the value mapped to e under the decompositon d

StructuredDecompositions.Decompositions.adhesions — Method

Get a vector of the adhesions of a decomposition

StructuredDecompositions.Decompositions.bags — Method

Get a vector of indexed bags; i.e. a vector consisting of pairs (x, dx) where x ∈ FG and dx is the value mapped to x under the decompositon d

StructuredDecompositions.Decompositions.bags — Method

Get a vector of the bags of a decomposition

StructuredDecompositions.Decompositions.∫ — Method

Syntactic sugar for costructing the category of elements of a graph. Note that ∫(G) has type Category whereas elements(G) has type Elements

StructuredDecompositions.Decompositions.𝐃 — Function

The construction of categories of structured decompostions is functorial; it consists of a functor 𝐃: Cat{pullback} → Cat taking any category C with pullbacks to the category 𝐃C of C-values structured decompositions. The functoriality of this construction allows us to lift any functor F : C → E to a functor 𝐃f : 𝐃 C → 𝐃 E which maps C-valued structured decompositions to E-valued structured decompositions. When we think of the functor F as a computational problem (taking inputs in C to solution spaces in E), then 𝐃f should be thought of as lifting the global comuputation of F to local computation on the constituent parts of C-valued decompositions. In particular, given a structured decomposition d: FG → C and a sheaf F: C → FinSet^{op} w.r.t to the decompositon topology, we can make a structured decomposition valued in FinSet^{op} by lifting the sheaf to a functor 𝐃f: 𝐃C → 𝐃(S^{op}) between categories of structured decompositions.

StructuredDecompositions.DecidingSheaves.decide_sheaf_tree_shape — Function

Solve the decision problem encoded by a sheaf. The algorithm is as follows: compute on each bag (optionally, if the decomposition of the solution space is already known, then it can be passed as an argument), compute composites on edges, project back down to bags answer (providing a witness) "no" if there is an empty bag; "yes" otherwise.

StructuredDecompositions.DecidingSheaves.old_adhesion_filter — Method

Filtering algorithm. Note: we are assuming that we only know how to work with FinSet(Int) !

INPUT: a Finset^{op}-valued structured decomposition d : FG → Span Finset^{op} (which is expected to be in co-decomposition form; i.e. as a diagram d : FG → Cospan Finset ) and an indexed span ( (ℓ, r), ( d(ℓ), d(r) ) ) in d (i.e a pair consisting of span (ℓ, r) in ∫G and its image under d)

OUTPUT: a structured decomposition obtained by replacing the span de in d by the span obtained by projecting the pullback of de (i.e. taking images)

StructuredDecompositions.DecidingSheaves.old_decide_sheaf_tree_shape — Function

Solve the decision problem encoded by a sheaf. The algorithm is as follows: compute on each bag (optionally, if the decomposition of the solution space is already known, then it can be passed as an argument), compute composites on edges, project back down to bags answer (providing a witness) "no" if there is an empty bag; "yes" otherwise.

StructuredDecompositions.JunctionTrees.AMDJL_AMD — Type

AMDJL_AMD <: EliminationAlgorithm

The approximate minimum degree algorithm. Uses AMD.jl.

StructuredDecompositions.JunctionTrees.AMDJL_SYMAMD — Type

AMDJL_SYMAMD <: EliminationAlgorithm

The SYMAMD algorithm. Uses AMD.jl.

StructuredDecompositions.JunctionTrees.CuthillMcKeeJL_RCM — Type

CuthillMcKeeJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses CuthillMckee.jl.

StructuredDecompositions.JunctionTrees.DFS — Type

DFS <: EliminationAlgorithm

The descent-first search algorithm.

StructuredDecompositions.JunctionTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

CuthillMcKeeJL_RCM

SymRCMJL_RCM

AMDJL_AMD

AMDJL_SYMAMD

MetisJL_ND

TreeWidthSolverJL_BT

MCS

DFS

StructuredDecompositions.JunctionTrees.Fundamental — Type

Fundamental <: Supernode

A fundamental supernode.

StructuredDecompositions.JunctionTrees.MCS — Type

MCS <: EliminationAlgorithm

The maximum cardinality search algorithm.

StructuredDecompositions.JunctionTrees.Maximal — Type

Maximal <: Supernode

A maximal supernode.

StructuredDecompositions.JunctionTrees.MetisJL_ND — Type

MetisJL_ND <: EliminationAlgorithm

The nested dissection heuristic. Uses Metis.jl.

StructuredDecompositions.JunctionTrees.Node — Type

Node <: Supernode

A single-vertex supernode.

StructuredDecompositions.JunctionTrees.Permutation — Type

Permutation <: AbstractVector{Int}

A permutation $\sigma$ of the set $\{1, \dots, n\}$. This type implements the abstract array interface.

StructuredDecompositions.JunctionTrees.Permutation — Method

Permutation(permutation::AbstractVector)

Construct a permutation $\sigma$ from a sequence $(\sigma(1), \dots, \sigma(n))$.

StructuredDecompositions.JunctionTrees.Permutation — Method

Permutation(graph[, ealg::EliminationAlgorithm])

Construct a fill-reducing permutation of the vertices of a graph.

StructuredDecompositions.JunctionTrees.SupernodeType — Type

SupernodeType

A type of supernode. The options are

Node

Maximal

Fundamental

StructuredDecompositions.JunctionTrees.SymRCMJL_RCM — Type

SymRCMJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses SymRCM.jl.

StructuredDecompositions.JunctionTrees.TreeWidthSolverJL_BT — Type

TreeWidthSolverJL_BT <: EliminationAlgorithm

The Bouchitte-Todinca algorithm. Uses TreeWidthSolver.jl.