JunctionTrees

StructuredDecompositions.JunctionTrees.Bag — Type

Bag <: AbstractVector{Int}

A bag of a junction tree.

StructuredDecompositions.JunctionTrees.JunctionTree — Type

JunctionTree <: AbstractVector{Bag}

A junction tree. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.junctiontree! — Function

junctiontree!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=AMDJL_AMD(),
    snd::SupernodeType=Maximal())

Construct a tree decomposition of a simple graph, represented by its adjacency matrix matrix. The vertices of the graph are first ordered by a fill-reducing permutation computed by the algorithm alg. The size of the resulting decomposition is determined by the supernode partition snd.

julia> using StructuredDecompositions.JunctionTrees

julia> graph = [
    0 1 1 0 0 0 0 0
    1 0 1 0 0 1 0 0
    1 1 0 1 1 0 0 0
    0 0 1 0 1 0 0 0
    0 0 1 1 0 0 1 1
    0 1 0 0 0 0 1 0
    0 0 0 0 1 1 0 1
    0 0 0 0 1 0 1 0
];

julia> label, tree = junctiontree(graph);

julia> tree
6-element JunctionTree:
[6, 7, 8]
├─ [1, 6, 7]
├─ [4, 6, 8]
│  └─ [3, 4, 6]
│     └─ [2, 3, 6]
└─ [5, 7, 8]

StructuredDecompositions.JunctionTrees.junctiontree — Function

junctiontree(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=AMDJL_AMD(),
    snd::SupernodeType=Maximal())

A non-mutating version of junctiontree!.

StructuredDecompositions.JunctionTrees.treewidth — Function

treewidth(tree::JunctionTree, i::Integer)

Compute the width of a junction tree.

StructuredDecompositions.JunctionTrees.separator — Function

separator(bag::Bag)

Get the separator of a bag.

separator(tree::JunctionTree, i::Integer)

Get the separator at node i.

StructuredDecompositions.JunctionTrees.residual — Function

residual(bag::Bag)

Get the residual of a bag.

residual(tree::JunctionTree, i::Integer)

Get the residual at node i.

Elimination Algorithms

StructuredDecompositions.JunctionTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

CuthillMcKeeJL_RCM

SymRCMJL_RCM

AMDJL_AMD

AMDJL_SymAMD

MetisJL_ND

TreeWidthSolverJL_BT

MCS

StructuredDecompositions.JunctionTrees.PermutationOrAlgorithm — Type

PermutationOrAlgorithm = Union{AbstractVector, EliminationAlgorithm}

Either a permutation or a graph elimination algorithm.

StructuredDecompositions.JunctionTrees.AMDJL_AMD — Type

AMDJL_AMD <: EliminationAlgorithm

The approximate minimum degree algorithm. Uses AMD.jl.

StructuredDecompositions.JunctionTrees.AMDJL_SymAMD — Type

AMDJL_SymAMD{T} <: EliminationAlgorithm

The SymAMD algorithm. Uses AMD.jl.

StructuredDecompositions.JunctionTrees.MetisJL_ND — Type

MetisJL_ND <: EliminationAlgorithm

The nested dissection heuristic. Uses Metis.jl.

StructuredDecompositions.JunctionTrees.CuthillMcKeeJL_RCM — Type

CuthillMcKeeJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses CuthillMckee.jl.

StructuredDecompositions.JunctionTrees.SymRCMJL_RCM — Type

SymRCMJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses SymRCM.jl.

StructuredDecompositions.JunctionTrees.TreeWidthSolverJL_BT — Type

TreeWidthSolverJL_BT <: EliminationAlgorithm

The Bouchitte-Todinca algorithm. Uses TreeWidthSolver.jl.

StructuredDecompositions.JunctionTrees.MCS — Type

MCS <: EliminationAlgorithm

The maximum cardinality search algorithm.

StructuredDecompositions.JunctionTrees.permutation — Function

permutation(matrix::AbstractMatrix, alg::PermutationOrAlgorithm)

Construct a fill-reducing permutation of the vertices of a simple graph, represented by its adjacency matrix matrix.

julia> using StructuredDecompositions.JunctionTrees

julia> graph = [
    0 1 1 0 0 0 0 0
    1 0 1 0 0 1 0 0
    1 1 0 1 1 0 0 0
    0 0 1 0 1 0 0 0
    0 0 1 1 0 0 1 1
    0 1 0 0 0 0 1 0
    0 0 0 0 1 1 0 1
    0 0 0 0 1 0 1 0
];

julia> order, index = permutation(graph, MCS());

julia> order
8-element Vector{Int64}:
 1
 6
 2
 3
 4
 5
 7
 8

Supernodes

StructuredDecompositions.JunctionTrees.SupernodeType — Type

SupernodeType

A type of supernode. The options are

Nodal

Maximal

Fundamental

StructuredDecompositions.JunctionTrees.Nodal — Type

Nodal <: Supernode

A single-vertex supernode.

StructuredDecompositions.JunctionTrees.Fundamental — Type

Fundamental <: Supernode

A fundamental supernode.

StructuredDecompositions.JunctionTrees.Maximal — Type

Maximal <: Supernode

A maximal supernode.