JunctionTrees

Junction Trees

StructuredDecompositions.JunctionTrees.Bag — Type

Bag <: AbstractVector{Int}

A bag of a junction tree.

StructuredDecompositions.JunctionTrees.JunctionTree — Type

JunctionTree <: AbstractVector{Bag}

A junction tree. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.junctiontree! — Function

junctiontree!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a tree decomposition of a simple graph. The vertices of the graph are first ordered by a fill-reducing permutation computed by the algorithm alg. The size of the resulting decomposition is determined by the supernode partition snd.

matrix: adjacency matrix

alg: ordering algorithm

snd: supernode partition type

julia> using StructuredDecompositions.JunctionTrees

julia> graph = [
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ];

julia> label, tree = junctiontree(graph);

julia> tree
6-element JunctionTree:
[6, 7, 8]
├─ [1, 6, 7]
├─ [4, 6, 8]
│  └─ [3, 4, 6]
│     └─ [2, 3, 6]
└─ [5, 7, 8]

StructuredDecompositions.JunctionTrees.junctiontree — Function

junctiontree(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

A non-mutating version of junctiontree!.

StructuredDecompositions.JunctionTrees.treewidth! — Function

treewidth!(matrix::SparseMatrixCSC;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Compute an upper bound to the tree width of a simple graph. See junctiontree! for the meaning of alg.

StructuredDecompositions.JunctionTrees.treewidth — Function

treewidth(tree::JunctionTree)

Compute the width of a junction tree.

treewidth(matrix::AbstractMatrix;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

A non-mutating version of treewidth!.

StructuredDecompositions.JunctionTrees.separator — Function

separator(bag::Bag)

Get the separator of a bag.

separator(tree::JunctionTree, i::Integer)

Get the separator at node i.

StructuredDecompositions.JunctionTrees.residual — Function

residual(bag::Bag)

Get the residual of a bag.

residual(tree::JunctionTree, i::Integer)

Get the residual at node i.

StructuredDecompositions.JunctionTrees.relative — Function

relative(tree::JunctionTree, i::Integer)

Get the indices in tree[parentindex(tree, i)] corresponding to the elements of separator(tree, i).

julia> using AbstractTrees

julia> using StructuredDecompositions.JunctionTrees

julia> graph = [
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ];

julia> label, tree = junctiontree(graph);

julia> bag = tree[parentindex(tree, 1)]
3-element Bag:
 6
 7
 8

julia> sep = separator(tree, 1)
2-element view(::Vector{Int64}, 1:2) with eltype Int64:
 6
 7

julia> rel = relative(tree, 1)
2-element view(::Vector{Int64}, 1:2) with eltype Int64:
 1
 2

julia> bag[rel] == sep
true

Chordal Graphs

StructuredDecompositions.JunctionTrees.chordalgraph — Function

chordalgraph([element=true,] tree::JunctionTree)

See below. The function returns a sparse matrix whose structural nonzeros are filled with element.

chordalgraph(Element::Type, tree::JunctionTree)

Construct the intersection graph of the subtrees of a junction tree. The function returns a sparse matrix with elements of type Element and the same sparsity structure as the lower triangular part of the graph's adjacency matrix.

StructuredDecompositions.JunctionTrees.ischordal — Function

ischordal(matrix::AbstractMatrix)

Determine whether a simple graph is chordal.

StructuredDecompositions.JunctionTrees.isperfect — Function

isperfect(matrix::AbstractMatrix, order::AbstractVector[, index::AbstractVector])

Determine whether an fill-reducing permutation is perfect.

Elimination Algorithms

StructuredDecompositions.JunctionTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

type	name	complexity

MCS	maximum cardinality search	O(m + n)

RCM	reverse Cuthill-Mckee	O(mΔ)

AMD	approximate minimum degree	O(mn)

SymAMD	column approximate minimum degree	O(mn)

MMD	multiple minimum degree	O(mn²)

NodeND	nested dissection

FlowCutter	FlowCutter

Spectral	spectral ordering

BT	Bouchitte-Todinca	O(2.6183ⁿ)

for a graph with m edges, n vertices, and maximum degree Δ.

StructuredDecompositions.JunctionTrees.PermutationOrAlgorithm — Type

PermutationOrAlgorithm = Union{AbstractVector, EliminationAlgorithm}

Either a permutation or an algorithm.

StructuredDecompositions.JunctionTrees.MCS — Type

MCS <: EliminationAlgorithm

MCS()

The maximum cardinality search algorithm.

StructuredDecompositions.JunctionTrees.RCM — Type

RCM <: EliminationAlgorithm

RCM(; sortbydeg=true)

The reverse Cuthill-McKee algorithm. Uses SymRCM.jl.

sortbydeg: whether to sort neighbor lists by degree

StructuredDecompositions.JunctionTrees.AMD — Type

AMD <: EliminationAlgorithm

AMD(; dense=nothing, aggressive=nothing)

The approximate minimum degree algorithm. Uses AMD.jl.

dense: dense row parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.SymAMD — Type

SymAMD{Index} <: EliminationAlgorithm

SymAMD{Index}(; dense_row=nothing, dense_col=nothing, aggressive=nothing) where Index

SymAMD(; dense_row=nothing, dense_col=nothing, aggressive=nothing)

The column approximate minimum degree algorithm. Uses AMD.jl.

Index: either Int or Cint

dense_row: dense row parameter

dense_column: dense column parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.MMD — Type

MMD <: EliminationAlgorithm

MMD()

The multiple minimum degree algorithm. Uses Sparspak.jl.

StructuredDecompositions.JunctionTrees.NodeND — Type

NodeND <: EliminationAlgorithm

NodeND()

The nested dissection algorithm. Uses Metis.jl.

StructuredDecompositions.JunctionTrees.FlowCutter — Type

FlowCutter <: EliminationAlgorithm

FlowCutter(; time=10, seed=0)

The FlowCutter algorithm. Uses FlowCutterPACE17_jll.jl.

time: running time

seed: random seed

StructuredDecompositions.JunctionTrees.Spectral — Type

Spectral <: EliminationAlgorithm

Spectral(; tol=0.0)

The spectral ordering algorithm. Uses Laplacians.jl.

tol: tolerance for convergence

StructuredDecompositions.JunctionTrees.BT — Type

BT <: EliminationAlgorithm

BT()

The Bouchitte-Todinca algorithm. Uses TreeWidthSolver.jl.

StructuredDecompositions.JunctionTrees.permutation — Function

permutation(matrix::AbstractMatrix, alg::PermutationOrAlgorithm)

Construct a fill-reducing permutation of the vertices of a simple graph.

julia> using StructuredDecompositions.JunctionTrees

julia> graph = [
    0 1 1 0 0 0 0 0
    1 0 1 0 0 1 0 0
    1 1 0 1 1 0 0 0
    0 0 1 0 1 0 0 0
    0 0 1 1 0 0 1 1
    0 1 0 0 0 0 1 0
    0 0 0 0 1 1 0 1
    0 0 0 0 1 0 1 0
];

julia> order, index = permutation(graph, MCS());

julia> order
8-element Vector{Int64}:
 1
 6
 2
 3
 4
 5
 7
 8

julia> index == invperm(order)
true

Supernodes

StructuredDecompositions.JunctionTrees.SupernodeType — Type

SupernodeType

A type of supernode partition. The options are

type	name

Nodal	nodal supernode partition

Maximal	maximal supernode partition

Fundamental	fundamental supernode partition

StructuredDecompositions.JunctionTrees.Nodal — Type

Nodal <: SupernodeType

A nodal supernode partition.

StructuredDecompositions.JunctionTrees.Fundamental — Type

Fundamental <: SupernodeType

A fundamental supernode partition.

StructuredDecompositions.JunctionTrees.Maximal — Type

Maximal <: SupernodeType

A maximal supernode partition.