JunctionTrees

Elimination Algorithms

StructuredDecompositions.JunctionTrees.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

type	name	complexity	connected

MCS	maximum cardinality search	O(m + n)	no

RCM	reverse Cuthill-Mckee	O(mΔ)	yes

AMD	approximate minimum degree	O(mn)	no

SymAMD	column approximate minimum degree		no

MMD	multiple minimum degree	O(mn²)	no

NodeND	nested dissection		no

Spectral	spectral ordering		yes

BT	Bouchitte-Todinca	O(2.6183ⁿ)	no

for a graph with m edges, n vertices, and maximum degree Δ. The algorithms RCM and Spectral only work on connected graphs.

StructuredDecompositions.JunctionTrees.PermutationOrAlgorithm — Type

PermutationOrAlgorithm = Union{AbstractVector, EliminationAlgorithm}

Either a permutation or an algorithm.

StructuredDecompositions.JunctionTrees.DEFAULT_ELIMINATION_ALGORITHM — Constant

DEFAULT_ELIMINATION_ALGORITHM = AMD()

The default algorithm.

StructuredDecompositions.JunctionTrees.MCS — Type

MCS <: EliminationAlgorithm

MCS()

The maximum cardinality search algorithm.

StructuredDecompositions.JunctionTrees.RCM — Type

RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm only works on connected graphs.

StructuredDecompositions.JunctionTrees.AMD — Type

AMD <: EliminationAlgorithm

AMD(; dense=10.0, aggressive=1.0)

The approximate minimum degree algorithm.

dense: dense row parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.SymAMD — Type

SymAMD <: EliminationAlgorithm

SymAMD(; dense_row=10.0, dense_col=10.0, aggressive=1.0)

The column approximate minimum degree algorithm.

dense_row: dense row parameter

dense_column: dense column parameter

aggressive: aggressive absorbtion

StructuredDecompositions.JunctionTrees.MMD — Type

MMD <: EliminationAlgorithm

MMD()

The multiple minimum degree algorithm.

StructuredDecompositions.JunctionTrees.NodeND — Type

NodeND <: EliminationAlgorithm

NodeND()

The nested dissection algorithm. In order to use it, import the package Metis.

StructuredDecompositions.JunctionTrees.Spectral — Type

Spectral <: EliminationAlgorithm

Spectral(; tol=0.0)

The spectral ordering algorithm only works on connected graphs. In order to use it, import the package Laplacians.

tol: tolerance for convergence

StructuredDecompositions.JunctionTrees.BT — Type

BT <: EliminationAlgorithm

BT()

The Bouchitte-Todinca algorithm. In order to use it, import the package TreeWidthSolver.

StructuredDecompositions.JunctionTrees.permutation — Function

permutation(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a fill-reducing permutation of the vertices of a simple graph.

julia> using SparseArrays, StructuredDecompositions

julia> graph = sparse([
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ]);

julia> order, index = permutation(graph);

julia> order
8-element Vector{Int64}:
 4
 8
 7
 6
 5
 1
 3
 2

julia> index == invperm(order)
true

Supernodes

StructuredDecompositions.JunctionTrees.SupernodeType — Type

SupernodeType

A type of supernode partition. The options are

type	name

Nodal	nodal supernode partition

Maximal	maximal supernode partition

Fundamental	fundamental supernode partition

StructuredDecompositions.JunctionTrees.DEFAULT_SUPERNODE_TYPE — Constant

DEFAULT_SUPERNODE_TYPE = Maximal()

The default supernode partition.

StructuredDecompositions.JunctionTrees.Nodal — Type

Nodal <: SupernodeType

A nodal supernode partition.

StructuredDecompositions.JunctionTrees.Maximal — Type

Maximal <: SupernodeType

A maximal supernode partition.

StructuredDecompositions.JunctionTrees.Fundamental — Type

Fundamental <: SupernodeType

A fundamental supernode partition.

Trees

StructuredDecompositions.JunctionTrees.AbstractTree — Type

AbstractTree{I} = Union{Tree{I}, SupernodeTree{I}, JunctionTree{I}}

A rooted forest. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.Tree — Type

Tree{V <: Integer} <: AbstractUnitRange{V}

Tree(tree::AbstractTree)

Tree{V}(tree::AbstractTree) where V

A rooted forest. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.SupernodeTree — Type

SupernodeTree{V} <: AbstractVector{UnitRange{V}}

A supernodal elimination tree. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.JunctionTree — Type

JunctionTree{V, E} <: AbstractVector{Bag{V, E}}

A junction tree. This type implements the indexed tree interface.

StructuredDecompositions.JunctionTrees.eliminationtree — Function

eliminationtree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Construct a tree-depth decomposition of a simple graph.

julia> using SparseArrays, StructuredDecompositions

julia> graph = sparse([
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ]);

julia> label, tree = eliminationtree(graph);

julia> tree
8-element Tree{Int64}:
8
└─ 7
   ├─ 5
   │  ├─ 1
   │  └─ 4
   │     └─ 3
   │        └─ 2
   └─ 6

StructuredDecompositions.JunctionTrees.supernodetree — Function

supernodetree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a supernodal elimination tree.

StructuredDecompositions.JunctionTrees.junctiontree — Function

junctiontree(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct a tree decomposition of a simple graph. The vertices of the graph are first ordered by a fill-reducing permutation computed by the algorithm alg. The size of the resulting decomposition is determined by the supernode partition snd.

julia> using SparseArrays, StructuredDecompositions

julia> graph = sparse([
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ]);

julia> label, tree = junctiontree(graph);

julia> tree
6-element JunctionTree{Int64}:
[6, 7, 8]
├─ [1, 6, 7]
├─ [4, 6, 8]
│  └─ [3, 4, 6]
│     └─ [2, 3, 6]
└─ [5, 7, 8]

StructuredDecompositions.JunctionTrees.treewidth — Function

treewidth(tree::JunctionTree)

Compute the width of a junction tree.

treewidth(graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM)

Compute an upper bound to the tree width of a simple graph. See junctiontree for the meaning of alg.

StructuredDecompositions.JunctionTrees.separator — Function

separator(bag::Bag)

Get the separator of a bag.

separator(tree::JunctionTree, i::Integer)

Get the separator at node i.

StructuredDecompositions.JunctionTrees.residual — Function

residual(bag::Bag)

Get the residual of a bag.

residual(tree::JunctionTree, i::Integer)

Get the residual at node i.

StructuredDecompositions.JunctionTrees.relative — Function

relative(tree::JunctionTree, i::Integer)

Get the relative indices at node i. These are not cached by default; compute them by running relative!(tree).

AbstractTrees.rootindex — Function

rootindex(tree::AbstractTree)

Get the first root of a rooted forest.

StructuredDecompositions.JunctionTrees.rootindices — Function

rootindices(tree::AbstractTree)

Construct an iterator over the roots of a rooted forest.

AbstractTrees.parentindex — Function

parentindex(tree::AbstractTree, i::Integer)

Get the parent index of node i. Returns nothing if i is a root.

StructuredDecompositions.JunctionTrees.firstchildindex — Function

firstchildindex(tree::AbstractTree, i::Integer)

Get the first child of node i. Returns nothing if i is a leaf.

AbstractTrees.nextsiblingindex — Function

nextsiblingindex(tree::AbstractTree, i::Integer)

Get the next sibling of node i. Returns nothing if i is the greatest sibling.

AbstractTrees.childindices — Function

childindices(tree::AbstractTree, i::Integer)

Construct an iterator over the children of node i.

Chordal Graphs

StructuredDecompositions.JunctionTrees.eliminationgraph — Function

eliminationgraph([element=true,] tree::JunctionTree)

See eliminationgraph!. The function returns a sparse matrix whose structural nonzeros are filled with element.

eliminationgraph([element=true,] graph;
    alg::PermutationOrAlgorithm=DEFAULT_ELIMINATION_ALGORITHM,
    snd::SupernodeType=DEFAULT_SUPERNODE_TYPE)

Construct the elimination graph of a simple graph.

julia> using SparseArrays, StructuredDecompositions

julia> graph = sparse([
           0 1 1 0 0 0 0 0
           1 0 1 0 0 1 0 0
           1 1 0 1 1 0 0 0
           0 0 1 0 1 0 0 0
           0 0 1 1 0 0 1 1
           0 1 0 0 0 0 1 0
           0 0 0 0 1 1 0 1
           0 0 0 0 1 0 1 0
       ]);

julia> label, filled = eliminationgraph(graph);

julia> filled
8×8 SparseMatrixCSC{Int64, Int64} with 13 stored entries:
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅
 1  1  1  1  ⋅  ⋅  ⋅  ⋅
 1  ⋅  ⋅  ⋅  1  1  ⋅  ⋅
 ⋅  ⋅  ⋅  1  1  1  1  ⋅

julia> isfilled(filled)
true

julia> ischordal(filled + filled')
true

StructuredDecompositions.JunctionTrees.eliminationgraph! — Function

eliminationgraph!(graph, tree::JunctionTree)

Construct the subtree graph of a junction tree. The result is stored in graph.

StructuredDecompositions.JunctionTrees.ischordal — Function

ischordal(graph)

Determine whether a simple graph is chordal.

StructuredDecompositions.JunctionTrees.isfilled — Function

isfilled(graph)

Determine whether a directed graph is filled.

StructuredDecompositions.JunctionTrees.isperfect — Function

isperfect(graph, order::AbstractVector[, index::AbstractVector])

Determine whether an fill-reducing permutation is perfect.